pyproject.toml
[tool.poetry]
name = "asset-downloader"
version = "0.1.0"
description = "My editing pal"
authors = ["Antonio Feregrino <antonio.feregrino@gmail.com>"]
license = "MIT"
packages = [
    { include = "asset_downloader" },
]

[tool.poetry.scripts]
adw = 'asset_downloader.__main__:app'

[tool.poetry.dependencies]
python = "^3.8"
pytube = "*"
typer = "*"
python-slugify = "*"
requests = "*"

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "^3.9.2"
isort = "^5.9.2"
pyproject-flake8 = "^0.0.1-alpha.2"
bump2version = "^1.0.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.venv          # root of the project
  )/
)
'''

[tool.flake8]
max-line-length = 120
per-file-ignores = [
]
exclude = [
    ".venv"
]

[tool.isort]
skip_glob = [
    ".venv"
]

# Structure
default_section = "THIRDPARTY"
known_first_party = "editingpal"

# Import Style
line_length = 120
force_grid_wrap = false
use_parentheses = true
include_trailing_comma = true
combine_as_imports = true
multi_line_output = 3
